+++ src/stores/index.ts
@@ -1,3 +1,4 @@
+export * from './aliasStore'
 export * from './envStore'
 export * from './gitStore'
 export * from './patchStore'